# docs/openapi/tenant-openapi.yaml
openapi: 3.1.0
info:
  title: CloudAPI Tenant
  description: '

    Welcome to CloudAPI Tenant!


    For detailed guidance on using the API, please visit our official documentation:

    https://www.didx.co.za/acapy-cloud/index.html

'
  version: 5.0.0-rc5
servers:
  - url: /tenant
paths:
...
  /v1/issuer/credentials:
    post:
      tags:
        - issuer
      summary: Send Holder a Credential
      description: "Create and send a credential, automating the issuer-side flow\n---\nNB: Only a tenant with the issuer role can send credentials.\n\nWhen sending a credential, the credential type must be one of `anoncreds` or `ld_proof`.\n```json\n{\n    \"anoncreds_credential_detail\": {...},\n    \"ld_credential_detail\": {...},\n    \"save_exchange_record\": true, <-- Whether the credential exchange record should be saved on completion.\n    \"connection_id\": \"string\", <-- The issuer's reference to the connection they want to submit the credential to.\n}\n```\nSetting the `save_exchange_record` field to True will save the exchange record after the credential is accepted.\nThe default behaviour is to only save exchange records while they are in pending state.\n\nFor a detailed technical specification of the credential issuing process, refer to the [Aries Issue Credential v2\nRFC](https://github.com/hyperledger/aries-rfcs/blob/main/features/0453-issue-credential-v2/README.md).\n\nRequest Body:\n---\n    credential: SendCredential\n        The payload for sending a credential\n\nReturns\n-------\n    CredentialExchange\n        A record of this credential exchange"
      operationId: send_credential_v1_issuer_credentials_post
      security:
        - APIKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCredential'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialExchange'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - issuer
      summary: Fetch Credential Exchange Records
      description: "Get a list of credential exchange records\n---\nBoth holders and issuers can call this endpoint, because they each have their own records of a credential exchange.\n\nThese records contain information about the credentials issued to a holder, such as the current state of the\nexchange, and other metadata such as the `connection_id` that a credential was submit to (if an issuer) or received\nfrom (if a holder). Each record in the list is related to a single credential exchange flow.\n\nNB: An issuer and a holder will have distinct credential exchange ids, despite referring to the same exchange.\nThe `thread_id` is the only record attribute that will be the same for the holder and the issuer.\n\nAn exchange record will automatically be deleted after a flow completes (i.e. when state is 'done'),\nunless the `save_exchange_record` was set to true.\n\nThe following parameters can be set to filter the fetched exchange records: connection_id, role, state, thread_id.\n\nParameters (Optional):\n---\n    limit: int - The maximum number of records to retrieve\n    offset: int - The offset to start retrieving records from\n    descending: bool - Whether to return results in descending order. Results are ordered by record created time.\n    connection_id: str\n    role: Role: \"issuer\", \"holder\"\n    state: State: \"proposal-sent\", \"proposal-received\", \"offer-sent\", \"offer-received\",\n                             \"request-sent\", \"request-received\", \"credential-issued\", \"credential-received\",\n                             \"credential-revoked\", \"abandoned\", \"done\"\n    thread_id: UUID\n\nReturns\n-------\n    List[CredentialExchange]\n        A list of credential exchange records"
      operationId: get_credentials_v1_issuer_credentials_get
      security:
        - APIKeyHeader: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
                maximum: 10000
                minimum: 1
              - type: 'null'
            description: Number of results to return
            default: 1000
            title: Limit
          description: Number of results to return
        - name: offset
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
                minimum: 0
              - type: 'null'
            description: Offset for pagination
            default: 0
            title: Offset
          description: Offset for pagination
        - name: descending
          in: query
          required: false
          schema:
            type: boolean
            description: Whether to return results in descending order. Results are ordered by record created time.
            default: true
            title: Descending
          description: Whether to return results in descending order. Results are ordered by record created time.
        - name: connection_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Connection Id
        - name: role
          in: query
          required: false
          schema:
            anyOf:
              - enum:
                  - issuer
                  - holder
                type: string
              - type: 'null'
            title: Role
        - name: state
          in: query
          required: false
          schema:
            anyOf:
              - enum:
                  - proposal-sent
                  - proposal-received
                  - offer-sent
                  - offer-received
                  - request-sent
                  - request-received
                  - credential-issued
                  - credential-received
                  - credential-revoked
                  - abandoned
                  - done
                  - deleted
                type: string
              - type: 'null'
            title: State
        - name: thread_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
                format: uuid
              - type: 'null'
            title: Thread Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CredentialExchange'
                title: Response Get Credentials V1 Issuer Credentials Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/issuer/credentials/create-offer:
    post:
      tags:
        - issuer
      summary: Create a Credential Offer (not bound to a connection)
      description: "Create a credential offer, not bound to any connection\n---\nNB: Only a tenant with the issuer role can send credentials.\n\nThis endpoint takes the same body as the send credential endpoint, but without a connection id. This\nmeans the credential will not be sent, but it will do the initial step of creating a credential exchange record,\nwhich the issuer can then use in the out of band (OOB) protocol.\n\nThe OOB protocol allows credentials to be sent over alternative channels, such as email or QR code, where a\nconnection does not yet exist between holder and issuer.\n\nThe credential type must be one of `anoncreds` or `ld_proof`.\n```json\n{\n    \"anoncreds_credential_detail\": {...},\n    \"ld_credential_detail\": {...},\n    \"save_exchange_record\": true, <-- Whether the credential exchange record should be saved on completion.\n}\n```\nFor a detailed technical specification of the credential issuing process, refer to the [Aries Issue Credential v2\nRFC](https://github.com/hyperledger/aries-rfcs/blob/main/features/0453-issue-credential-v2/README.md).\n\nRequest Body:\n---\n    credential: CreateOffer\n        The payload for creating a credential offer\n\nReturns\n-------\n    CredentialExchange\n        A record of this credential exchange"
      operationId: create_offer_v1_issuer_credentials_create_offer_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOffer'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialExchange'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - APIKeyHeader: []
  /v1/issuer/credentials/{credential_exchange_id}/request:
    post:
      tags:
        - issuer
      summary: Accept a Credential Offer
      description: "Sends a request to accept a credential offer\n---\nThe holder uses this endpoint to accept an offer from an issuer.\n\nIn technical terms, when a holder has a credential exchange record with a state 'offer-received', then they can use\nthis endpoint to accept that credential offer, and store the credential in their wallet.\n\nParameters\n----------\n    credential_exchange_id: str\n        The holder's reference to the credential exchange that they want to accept\n    save_exchange_record: Optional[bool]\n        Whether to override environment setting for saving credential exchange records. Default is None (use\n        environment setting). True means save record, False means delete record.\n\nReturns\n-------\n    CredentialExchange\n        An updated record of this credential exchange"
      operationId: request_credential_v1_issuer_credentials__credential_exchange_id__request_post
      security:
        - APIKeyHeader: []
      parameters:
        - name: credential_exchange_id
          in: path
          required: true
          schema:
            type: string
            title: Credential Exchange Id
        - name: save_exchange_record
          in: query
          required: false
          schema:
            anyOf:
              - type: boolean
              - type: 'null'
            description: Controls exchange record retention after exchange is complete. None uses wallet default (typically to delete), true forces save, false forces delete.
            title: Save Exchange Record
          description: Controls exchange record retention after exchange is complete. None uses wallet default (typically to delete), true forces save, false forces delete.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialExchange'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/issuer/credentials/{credential_exchange_id}/store:
    post:
      tags:
        - issuer
      summary: Store a Received Credential in Wallet
      description: "Store a received credential in wallet\n---\nStore a received credential in wallet by providing the credential exchange id.\n\n`credential_exchange_id` is the id of the exchange record, not the credential itself.\n\nThe holder only needs to call this endpoint if the holder receives a credential out of band, since, typically,\ncredentials are automatically stored in wallet after they are accepted.\n\nThe holder can store the credential in their wallet after receiving it from the issuer.\n\nParameters\n----------\n    credential_exchange_id: str\n        credential exchange record identifier\n\nReturns\n-------\n    CredentialExchange\n        An updated record of this credential exchange"
      operationId: store_credential_v1_issuer_credentials__credential_exchange_id__store_post
      security:
        - APIKeyHeader: []
      parameters:
        - name: credential_exchange_id
          in: path
          required: true
          schema:
            type: string
            title: Credential Exchange Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialExchange'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AnonCredsCredential:
      properties:
        credential_definition_id:
          type: string
          title: Credential Definition Id
        issuer_did:
          anyOf:
            - type: string
            - type: 'null'
          title: Issuer Did
          description: The DID to use as the issuer of the credential. If not provided, the public DID of the issuer wallet will be used.
        attributes:
          additionalProperties:
            type: string
          type: object
          title: Attributes
      type: object
      required:
        - credential_definition_id
        - attributes
      title: AnonCredsCredential
...
    CreateOffer:
      properties:
        save_exchange_record:
          anyOf:
            - type: boolean
            - type: 'null'
          title: Save Exchange Record
          description: Controls exchange record retention after exchange is complete. None uses wallet default (typically to delete), true forces save, false forces delete.
        ld_credential_detail:
          anyOf:
            - $ref: '#/components/schemas/LDProofVCDetail'
            - type: 'null'
        anoncreds_credential_detail:
          anyOf:
            - $ref: '#/components/schemas/AnonCredsCredential'
            - type: 'null'
      type: object
      title: CreateOffer
...
    CredentialExchange:
      properties:
        attributes:
          anyOf:
            - additionalProperties:
                type: string
              type: object
            - type: 'null'
          title: Attributes
        connection_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Connection Id
        created_at:
          type: string
          title: Created At
        credential_definition_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Credential Definition Id
        credential_exchange_id:
          type: string
          title: Credential Exchange Id
        did:
          anyOf:
            - type: string
            - type: 'null'
          title: Did
        error_msg:
          anyOf:
            - type: string
            - type: 'null'
          title: Error Msg
        role:
          type: string
          enum:
            - issuer
            - holder
          title: Role
        schema_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Schema Id
        state:
          anyOf:
            - type: string
              enum:
                - proposal-sent
                - proposal-received
                - offer-sent
                - offer-received
                - request-sent
                - request-received
                - credential-issued
                - credential-received
                - credential-revoked
                - abandoned
                - done
                - deleted
            - type: 'null'
          title: State
        thread_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Thread Id
        type:
          type: string
          title: Type
          default: anoncreds
        updated_at:
          type: string
          title: Updated At
      type: object
      required:
        - created_at
        - credential_exchange_id
        - role
        - updated_at
      title: CredentialExchange
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LDProofVCDetail:
      properties:
        credential:
          $ref: '#/components/schemas/Credential'
          description: Detail of the JSON-LD Credential to be issued
        options:
          $ref: '#/components/schemas/LDProofVCOptions'
          description: Options for specifying how the linked data proof is created.
        additional_properties:
          additionalProperties: true
          type: object
          title: Additional Properties
          default: {}
      type: object
      required:
        - credential
        - options
      title: LDProofVCDetail
      description: LDProofVCDetail
...
    SendCredential:
      properties:
        save_exchange_record:
          anyOf:
            - type: boolean
            - type: 'null'
          title: Save Exchange Record
          description: Controls exchange record retention after exchange is complete. None uses wallet default (typically to delete), true forces save, false forces delete.
        ld_credential_detail:
          anyOf:
            - $ref: '#/components/schemas/LDProofVCDetail'
            - type: 'null'
        anoncreds_credential_detail:
          anyOf:
            - $ref: '#/components/schemas/AnonCredsCredential'
            - type: 'null'
        connection_id:
          type: string
          title: Connection Id
      type: object
      required:
        - connection_id
      title: SendCredential
...
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: x-api-key